Initial debdeploy prototype
Initial prototype of debdeploy. It can deploy/rollback updates, detects
necessary process restarts after the update of a library/runtime interpreter
and allows central service restarts for server groups.

This is an initial snapshot to be used for updating a few servers as
test systems to give it some real world testing.

All WIP, some features such as the central package database
are not yet done/integrated.