Remove debdeploy postinsts
Drop the automatic restart of the salt master in the postinst of
debdeploy-master. Since the restart is rather intrusive (all minions
need to reconnect, it should rather be done explicitly at an appropriate
time.

Also drop the reload of the salt minion in postinst. This should rather be
done via "salt '*' saltutil.refresh_modules". Having it done in the postinst
breaks the update of debdeploy via debdeploy/salt on systemd-based systems.

Change-Id: Ida3ac3e6cbffc4aa73095ef3be7aa75115cdf768