Readd rollback handling to debdeploy
With debdeploy/salt rollbacks relied on the job database managed by Salt.
That no longer exists with Cumin, so instead it uses the following scheme:

To rollback an update, the original YAML update spec needs to be passed
to rollback-update(1), which queries the versions to downgrade to. This
new spec file can be then used like any other deployment (except that
it enables a new flag with enables downgrades internally)

Change-Id: I20738477f580d79888c805d6ff369c4267b99c00