debdeploy: Change the `--servers` flag to a global flag
The --servers flag is used by all subparsers so to be more DRY it
should be moved to the root parser.  This does change the way one
calls this script from:
  `sudo debdeploy deploy --update 2019-10-03-openssl.yaml --servers kubernetes-all`
to:
  `sudo debdeploy --servers kubernetes-all deploy --update 2019-10-03-openssl.yaml`
which means muscle memory will need to be updated.  However it will make
it easier to add additional query support

Change-Id: I8ebd2c089c8c62d9d398ed371170b637c4d7fd6c