Export DEBIAN_FRONTEND=noninteractive in the debdeploy client
There's a long-standing cornercase/bug in the debdeploy client which can
lead to hanging updates if an update triggers a debconf prompt. This is
fairly rare, but sometimes happens if we've puppetised a config file which is
also modified in a Debian postinst.

Conffile works the following way in dpkg:

1. Each file marked as a conffile records a hash value of it's pristine
form (as shipped in the package)
2. If a packages is updated and the new package ships an updated version
of a config file (e.g. because default values changed), dpkg checks the
old hash value. If it's matches the current file, the user made no local
modifications and it can be expected that she also wants the default settings
for the updated version.
3. Enter Puppet; it overrides the default settings shipped in the deb package
4. When running apt-get from the terminal this triggers a prompt which let
the user diff the modified settings and the new incoming one and select whether
to retain the local version or accept the new default.

As we rely on Puppet we usually want the puppetised value to be kept (hopefully
we've synched the changes to the default setting to the file distributed via
Puppet!). We were already passing --force-confold and --force-confdef to
that effect, but when run remotely via Cumin the debconf dialogue was still
shown, which effectively led to the Cumin process wait on the (never happening)
user input.

We can however prevent the display of the debconf dialogue by exporting
DEBIAN_FRONTEND=noninteractive. I've tested this successfully on cp2021 with
an openssh update (which triggers exactly the bug described above).

Change-Id: I9fa31640ff73c82fde3e22910d6bb1e5ad1a5341